+++ src/app/app.component.ts
@@ -65,9 +65,9 @@ export class AppComponent implements OnInit, OnDestroy {
      * @param {TrackingService} trackingService
      */
     public constructor(
-        private router: Router,
+        private readonly router: Router,
         private titleService: TitleService,
-        private trackingService: TrackingService,
+        private readonly trackingService: TrackingService,
     ) {}
 
     public ngOnInit(): void {
@@ -80,8 +80,8 @@ export class AppComponent implements OnInit, OnDestroy {
             )
             .subscribe((event: NavigationEnd) => {
                 this.show()
-                this.trackingService.track('pageview', {
-                    page: event.urlAfterRedirects,
+                this.trackingService.track('page_view', {
+                    page_path: event.urlAfterRedirects,
                 })
             })
     }