+++ src/app/components/page-header/page-header.component.ts
@@ -11,6 +11,4 @@ export class PageHeaderComponent {
     @Input() public title: string
 
     @Input() public copy: string
-
-    public constructor() {}
 }